Types: fix strictness and add a failing test
Because this library used the ember-cli-typescript generator to
configure TypeScript initially, it did not set `strict: true` in the
`tsconfig.json`, and the generator specifically does not enable the
`strictFunctionTypes` flag because the types for Ember's `get` and `set`
functions must be covariant to work with Ember's mixin system.

The result was that we missed a very important case for TypeScript
consumers of the package, specifically with function-based modifiers:
users could only pass `Element` as the type of the first argument to the
callback for the modifier. They *should* be able to pass *any* subtype
of element there.

This commit adds a failing test for this case, as well as a test to
ensure that the narrowing behavior works as expected for classes (which
it already does, because class properties have different assignability
rules than callbacks, courtesy of variance).