Types: make functional modifier contravariance-safe
Following on from the previous commit's tsconfig fix and new type test,
implement the bug fix so that the contravariant-position argument works
as desired. In less-jargony terms, this was previously a type error,
as it allowed *only* `el: Element`:

    export default modifier((el: HTMLDivElement) => {
      console.log(el.tagName);
    })

Now it correctly allows this and is *not* a type error.